Nicify messages, add ignore option, print failed and errored tests

+++ src/test.rs
@@ -12,6 +12,13 @@ pub struct Test {
     expected: PathBuf,
     cmd: String,
     quiet: bool,
+    ignore: bool,
+}
+
+pub enum TestResult {
+    Success,
+    Ignored,
+    Fail,
 }
 
 impl Test {
@@ -20,7 +27,8 @@ impl Test {
                expected: &str,
                command: &str,
                directory: &::std::fs::DirEntry,
-               quiet: bool)
+               quiet: bool,
+               ignore: bool)
                -> Test {
         let name = directory.file_name();
         let directory = directory.path().to_path_buf();
@@ -40,10 +48,15 @@ impl Test {
             expected: directory.join(expected),
             cmd: cmd,
             quiet: quiet,
+            ignore: ignore,
         }
     }
 
-    pub fn run(&self) -> Result<bool, io::Error> {
+    pub fn run(&self) -> Result<TestResult, io::Error> {
+        if self.ignore && self.name.starts_with("ignore") {
+            return Ok(TestResult::Ignored);
+        }
+
         // Prepare temp
         let _ = fs::remove_dir_all(&self.temp);
         filesystem::recursive_copy(&self.source, &self.temp)?;
@@ -62,8 +75,10 @@ impl Test {
         let ans = filesystem::compare_dirs(&self.temp, &self.expected).unwrap_or(false);
         if ans {
             fs::remove_dir_all(&self.temp)?;
+            Ok(TestResult::Success)
+        } else {
+            Ok(TestResult::Fail)
         }
-        Ok(ans)
     }
 
     pub fn name(&self) -> &str {
